Tag list now shows only unique tags with the count of the tags currently in use.  When selecting a tag link the user is brought to the tag usage index and there the user has a rough-draft view of all the scenarios/features that currently employ the tag.
Tags on the feature pages will link to the tag index.

Updated the Tag object to contain references to a feature and scenario to make it easier to trackback to names of objects for linking.

Created a TagUage object within the template for the purposes of using solely for the templates to work correclty with my array of Tags.  I don't believe this is correct use and again this after parsing manipulation of the data seems against the layout created by YARD.