Matching Step Definition and Step Transform Optimizations
Started reviewing the code to begin optimization process and found
quite a bit of waste around continually grabbing all the step definitions
and step transforms from the registry, iterating, and creating
regular expression to make the matches.

Created a cache of the definitions and transforms in a hash with the
comparison regex as the key.  Makes the compare a much faster process.

Also make transform look ups at the same time while in the middle of
the match instead of in a separate transaction.