Feature Handler Cache of Steps Transformers
Instead of creating a new cache of step transform and step definition
each time a feature is called it is created once and stored in a member
variable.  Since I control the order of the processing and made features
last in the processing order this will totally work well to keep the
system optimized.